Moved table.hpp to include/. Split off src/table/table.cpp.
