+++ cpp/tests/copying/gather_tests.cu
@@ -15,7 +15,7 @@
  */
 
 #include "copying.hpp"
-#include <table/table.hpp>
+#include <table.hpp>
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
